add graphQL Schema in schema file
