+++ .meteor/packages
@@ -18,3 +18,4 @@ typescript@4.2.2              # Enable TypeScript syntax in .ts and .tsx modules
 shell-server@0.5.0            # Server-side component of the `meteor shell` command
 
 apollo@2.0.0
+swydo:graphql
