+++ .meteor/versions
@@ -67,6 +67,7 @@ spacebars-compiler@1.3.0
 standard-minifier-css@1.7.3
 standard-minifier-js@2.6.1
 static-html@1.3.2
+swydo:graphql@1.0.0
 templating-tools@1.2.1
 tracker@1.2.0
 typescript@4.2.2
