+++ imports/ui/App.js
@@ -3,13 +3,13 @@ import gql from "graphql-tag";
 import { graphql } from "react-apollo";
 
 const App = ({ data }) => {
-  return <h1>Hello {data.greet}</h1>;
+  return <h1>Hello, {data.greet}</h1>;
 };
 
-const greetQuery = gql`
+const query = gql`
   {
     greet
   }
 `;
 
-export default graphql(greetQuery)(App);
+export default graphql(query)(App);
